move some errors from scylla_cql to scylla
