+++ scylla/src/tests/session_test.rs
@@ -2398,7 +2398,7 @@ async fn test_rate_limit_exceeded_exception() {
         }
     }
 
-    use scylla_cql::errors::OperationType;
+    use crate::execution::errors::OperationType;
 
     match maybe_err.expect("Rate limit error didn't occur") {
         QueryError::DbError(DbError::RateLimitReached { op_type, .. }, _) => {
